last commit made floating platforms work. must be an edge collider for floating platform and a box collider for the main ground.
made this work by adding 2 checks to turn on and off the feetBox which is a child of the player(feetBox is what stops the player from falling through the floor)

OnTriggerExit2D didn't work so I need to make it have two checks. not the way I'd like it to work but it does for now.

This commit just commented out the debug.log shit saying that the checks for the feetBox on and off were working.